Added `hitValue` and generic typing to `Polyline`s, and reflected change elsewhere Removed polyline simplification precomputer Reorganised polyline-related source files Fixed bugs and improved efficiency
