+++ test/test_utils/test_app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_map/src/layer/circle_layer.dart';
 import 'package:flutter_map/src/layer/marker_layer.dart';
 import 'package:flutter_map/src/layer/polygon_layer/polygon_layer.dart';
-import 'package:flutter_map/src/layer/polyline_layer.dart';
+import 'package:flutter_map/src/layer/polyline_layer/polyline_layer.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_layer.dart';
 import 'package:flutter_map/src/map/controller/map_controller.dart';
 import 'package:flutter_map/src/map/options/options.dart';
